Separate slave logs, and master logs.
